webapp/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-multipart==0.0.9
pydub==0.25.1
librosa==0.10.1
numpy==1.26.4
joblib==1.4.2
scikit-learn==1.4.2
python-dotenv==1.0.1
aiofiles==23.2.1
soundfile==0.12.1
